Remove unused python code.
Now that the bulk has transferred we can make a sweeping delete of the
old directory structures.  There is still reporting and executing so we
can't remove all of python yet, but this makes the remainders easier to
spot.  It also removes the redundancies provided by Haskell out of the
box.