Drop python 3.6 support in this version.
With the use of dataclasses we no longer support python 3.6.  I'm
explicitly avoiding the compatibility library for two reasons.  One, I
can't find a nice way to use it only with python 3.6 in poetry.  TWo,
this isn't a library as much as a binary executable that we can make
decisions like this.  If for some reason this breaks a flow for someone
somewhere then we can always remove this commit and keep it until 3.6
EOLs.