Fix compress option to include off.
This not only fixes the issue reported in #130 but also ensures that the
conversion to two strings remains total if new values are added in the
future.  The Enum click conversion fixes were unexpected but seem to be
due to the default values getting passed in as the type rather than the
string values one would expect.  I'll have a follow up diff that ensures
defaults are strings and fix the typing.