Update devcontainer configuration.
Some minor changes have happened in the VS Code devcontainers configuration.  I
can no longer get this to start with the git feature (even the new replacement
feature).  This allows the environment to start so we can continue making
progress with this project, but a deeper study of devcontainer environments will
be needed at some point.